--- admin-jobs/app/controllers/HealthCheck.scala
@@ -1,8 +1,0 @@
-package controllers
-
-import conf.{AllGoodCachedHealthCheck, NeverExpiresSingleHealthCheck}
-import play.api.libs.ws.WSClient
-
-class HealthCheck(wsClient: WSClient) extends AllGoodCachedHealthCheck(
-  NeverExpiresSingleHealthCheck("/news-alert/alerts")
-)(wsClient)
