--- admin-jobs/conf/application.conf
@@ -1,44 +1,0 @@
-
-akka {
-    akka.loggers = ["akka.event.Logging$DefaultLogger", "akka.event.slf4j.Slf4jLogger"]
-    loglevel = WARNING
-    actor {
-        default-dispatcher = {
-            fork-join-executor {
-                parallelism-factor = 1.0
-                parallelism-max = 24
-            }
-        }
-        java-futures = {
-            fork-join-executor {
-                parallelism-factor = 1.0
-                parallelism-max = 1
-            }
-        }
-    }
-}
-
-play {
-
-    crypto {
-        # The secret key is used to secure cryptographics functions.
-        # If you deploy your application to several instances be sure to use the same key!
-        secret: ${APP_SECRET}
-    }
-
-    il8n {
-        langs: "en"
-    }
-
-    ws {
-        compressionEnabled: true
-    }
-
-    http.forwarded.trustedProxies = [ "0.0.0.0/0" ]
-
-    application.loader = AppLoader
-}
-
-guardian: {
-    projectName: admin-jobs
-}
