--- admin-jobs/conf/logback.xml
@@ -1,22 +1,0 @@
-<configuration>
-
-    <contextName>frontend-admin-jobs</contextName>
-
-    <appender name="LOGFILE" class="ch.qos.logback.core.rolling.RollingFileAppender">
-        <file>logs/frontend-admin-jobs.log</file>
-
-        <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-            <fileNamePattern>logs/frontend-admin-jobs.log.%d{yyyy-MM-dd}.gz</fileNamePattern>
-            <maxHistory>7</maxHistory>
-        </rollingPolicy>
-
-        <encoder>
-            <pattern>%date [%thread] %-5level %logger{36} - %msg%n%xException</pattern>
-        </encoder>
-    </appender>
-
-    <root level="INFO">
-        <appender-ref ref="LOGFILE"/>
-    </root>
-
-</configuration>
