--- admin-jobs/conf/routes
@@ -1,8 +1,0 @@
-# Routes
-# This file defines all application routes (Higher priority routes first)
-# ~~~~
-
-GET     /_healthcheck                                                           controllers.HealthCheck.healthCheck()
-
-GET     /news-alert/alerts                                                      controllers.NewsAlertController.alerts()
-POST    /news-alert/alert                                                       controllers.NewsAlertController.create()
