+++ onward/app/models/BreakingNews.scala
@@ -2,6 +2,7 @@ package models
 
 import java.net.URI
 import java.util.UUID
+
 import org.joda.time.DateTime
 import play.api.libs.functional.syntax._
 import play.api.libs.json._
