+++ riff-raff.yaml
@@ -14,8 +14,6 @@ templates:
 deployments:
   admin:
     template: frontend
-  admin-jobs:
-    template: frontend
   applications:
     template: frontend
   archive:
